fix production db

+++ db/migrate/20141012182950_change_data_type_for_phone.rb
@@ -1,14 +1,14 @@
 class ChangeDataTypeForPhone < ActiveRecord::Migration
   def self.up
     change_table :messages do |t|
-      t.change :phone, :double
-      t.change :cellphone, :double
+      t.change :phone, :float
+      t.change :cellphone, :float
     end
   end
   def self.down
     change_table :messages do |t|
-      t.change :phone, :double
-      t.change :cellphone, :double
+      t.change :phone, :integer
+      t.change :cellphone, :integer
     end
   end
 end
